+++ Arbiter.App/Converters/PacketDirectionColorConverter.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.Globalization;
 using Arbiter.App.Models;
+using Arbiter.App.ViewModels.Tracing;
 using Arbiter.Net;
 using Arbiter.Net.Client;
+using Arbiter.Net.Filters;
 using Arbiter.Net.Server;
 using Avalonia.Data.Converters;
 using Avalonia.Media;
@@ -12,25 +14,46 @@ namespace Arbiter.App.Converters;
 public class PacketDirectionColorConverter : IValueConverter
 {
     private static readonly SolidColorBrush DefaultBrush = new(Color.FromRgb(0xcd, 0xcf, 0xd2));
+
     private static readonly SolidColorBrush ClientBrush = new(Color.FromRgb(0x82, 0xbf, 0xf7));
+    private static readonly SolidColorBrush ClientReplaceBrush = new(Color.FromRgb(0x82, 0xbf, 0xf7));
+
     private static readonly SolidColorBrush ServerBrush = new(Color.FromRgb(0x83, 0xe6, 0x7a));
+    private static readonly SolidColorBrush ServerReplaceBrush = new(Color.FromRgb(0x83, 0xe6, 0x7a));
     
+    private static readonly SolidColorBrush BlockedBrush = new(Color.FromRgb(0x71, 0x74, 0x79));
+        
     public static PacketDirectionColorConverter Converter => new();
-    
+
     public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
     {
+        var action = NetworkFilterAction.Allow;
+
+        // If value is a TracePacketViewModel, use the filter action and direction from it
+        if (value is TracePacketViewModel vm)
+        {
+            action = vm.FilterAction;
+
+            if (action == NetworkFilterAction.Block)
+            {
+                return BlockedBrush;
+            }
+
+            value = vm.Direction;
+        }
+        
         return value switch
         {
             PacketDirection => value switch
             {
-                PacketDirection.Client => ClientBrush,
-                PacketDirection.Server => ServerBrush,
+                PacketDirection.Client => action == NetworkFilterAction.Replace ? ClientReplaceBrush : ClientBrush,
+                PacketDirection.Server => action == NetworkFilterAction.Replace ? ServerReplaceBrush : ServerBrush,
                 _ => DefaultBrush
             },
             NetworkPacket => value switch
             {
-                ClientPacket => ClientBrush,
-                ServerPacket => ServerBrush,
+                ClientPacket => action == NetworkFilterAction.Replace ? ClientReplaceBrush : ClientBrush,
+                ServerPacket => action == NetworkFilterAction.Replace ? ServerReplaceBrush : ServerBrush,
                 _ => DefaultBrush
             },
             _ => null